+++ packages/core/src/builder/transactionBuilder.ts
@@ -286,6 +286,15 @@ export class TransactionBuilder {
     let inputs = selector.select(target);
 
     if (isDefined(this._changeAddress)) {
+      const firstInputId = inputs[0].boxId;
+      const manualMintingTokenId = target.tokens.some((x) => x.tokenId === firstInputId)
+        ? firstInputId
+        : undefined;
+
+      if (manualMintingTokenId) {
+        target.tokens = target.tokens.filter((x) => x.tokenId !== manualMintingTokenId);
+      }
+
       let change = utxoDiff(utxoSum(inputs), target);
       const changeBoxes: OutputBuilder[] = [];
 
@@ -363,10 +372,7 @@ export class TransactionBuilder {
     }
 
     if (some(burning.tokens) && some(this._burning)) {
-      burning = utxoDiff(burning, {
-        nanoErgs: _0n,
-        tokens: this._burning.toArray()
-      });
+      burning = utxoDiff(burning, { nanoErgs: _0n, tokens: this._burning.toArray() });
     }
 
     if (!this._settings.canBurnTokens && some(burning.tokens)) {
@@ -376,16 +382,27 @@ export class TransactionBuilder {
     return unsignedTransaction;
   }
 
+  private _getMintingOutput(): OutputBuilder | undefined {
+    for (const output of this._outputs) {
+      if (output.minting) return output;
+    }
+
+    return;
+  }
+
   private _isMinting(): boolean {
-    for (const output of this._outputs) if (output.minting) return true;
-    return false;
+    return this._getMintingOutput() !== undefined;
+  }
+
+  private _getMintingTokenId(): string | undefined {
+    return this._getMintingOutput()?.minting?.tokenId;
   }
 
   private _isMoreThanOneTokenBeingMinted(): boolean {
     let mintingCount = 0;
 
     for (const output of this._outputs) {
-      if (isDefined(output.minting)) {
+      if (output.minting) {
         mintingCount++;
         if (mintingCount > 1) return true;
       }
@@ -407,18 +424,6 @@ export class TransactionBuilder {
     }
 
     return false;
-  }
-
-  private _getMintingTokenId(): string | undefined {
-    let tokenId = undefined;
-    for (const output of this._outputs) {
-      if (output.minting) {
-        tokenId = output.minting.tokenId;
-        break;
-      }
-    }
-
-    return tokenId;
   }
 }